Quieten apt-get a little to remove the dynamic progress output
I was hoping for an explicit config option for this, however there
doesn't seem to be one. The man page for apt-get notes that the
first level of --quiet (there are several) just turns off the
progress output to make apt-get more suitable for logging.

This is mainly intended to make the output on CircleCI somewhat
easier to read. Ideally we'd have a way to switch this off when
running the build locally (where it's nice to have the pretty
output), though that is a rarer scenario so I'm not particularly
concerned.